Renaming project to NomaiGrandPrix and updating credits
