+++ NomaiGrandPrix.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30503.244
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OuterWildsRandomSpeedrun", "OuterWildsRandomSpeedrun\OuterWildsRandomSpeedrun.csproj", "{368A7F4A-8AEF-426B-992A-31B6CF7A7ADA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NomaiGrandPrix", "NomaiGrandPrix\NomaiGrandPrix.csproj", "{368A7F4A-8AEF-426B-992A-31B6CF7A7ADA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
