+++ NomaiGrandPrix/Constants.cs
@@ -1,0 +1,13 @@
+using UnityEngine;
+
+namespace NomaiGrandPrix
+{
+  public struct Constants
+  {
+    public static Color OW_ORANGE_COLOR = new Color(0.968f, 0.498f, 0.207f);
+    public static Color OW_SELECTED_COLOR = new Color(0.9882f, 0.8627f, 0.7686f);
+    public const string OW_MENU_FONT_NAME = "Adobe - SerifGothicStd-ExtraBold";
+    public const string SPEEDRUN_BUTTON_TEXT = "NOMAI GRAND PRIX";
+    public const string RESET_RUN_BUTTON_TEXT = "RANDOMIZE SPAWN AND GOAL";
+  }
+}
