+++ NomaiGrandPrix/NomaiGrandPrix.csproj.user
@@ -1,0 +1,5 @@
+﻿<Project>
+  <PropertyGroup>
+    <OutputPath>$(AppData)\OuterWildsModManager\OWML\Mods\NomaiGrandPrix</OutputPath>
+  </PropertyGroup>
+</Project>
