+++ NomaiGrandPrix/SpawnPointPool.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 
-namespace OuterWildsRandomSpeedrun
+namespace NomaiGrandPrix
 {
     public enum Area
     {
