+++ NomaiGrandPrix/manifest.json
@@ -1,0 +1,10 @@
+{
+  "$schema": "https://raw.githubusercontent.com/ow-mods/owml/master/schemas/manifest_schema.json",
+  "filename": "NomaiGrandPrix.dll",
+  "author": "Zac Bauermeister, David Damschen, Ben Kirzhner, Chris Maines",
+  "name": "Nomai Grand Prix",
+  "uniqueName": "NomaiGrandPrix",
+  "version": "0.1.0",
+  "owmlVersion": "2.9.0",
+  "dependencies": []
+}
