--- OuterWildsRandomSpeedrun/OuterWildsRandomSpeedrun.csproj.user
@@ -1,5 +1,0 @@
-﻿<Project>
-  <PropertyGroup>
-    <OutputPath>$(AppData)\OuterWildsModManager\OWML\Mods\ZacBauermeister.OuterWildsRandomSpeedrun</OutputPath>
-  </PropertyGroup>
-</Project>
