--- OuterWildsRandomSpeedrun/manifest.json
@@ -1,10 +1,0 @@
-{
-  "$schema": "https://raw.githubusercontent.com/ow-mods/owml/master/schemas/manifest_schema.json",
-  "filename": "OuterWildsRandomSpeedrun.dll",
-  "author": "Zac Bauermeister",
-  "name": "OuterWildsRandomSpeedrun",
-  "uniqueName": "ZacBauermeister.OuterWildsRandomSpeedrun",
-  "version": "0.1.0",
-  "owmlVersion": "2.9.0",
-  "dependencies": []
-}
